backend/requirements.txt
Django==4.2.7
channels==4.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
python-docx==1.0.1
PyPDF2==3.0.1
bcrypt==4.0.1
django-oauth-toolkit==2.3.0
drf-yasg==1.21.7
Pillow==10.1.0
python-jose==3.3.0
websockets==12.0
